chore(aria-required-parent): work with standards object
