+++ lib/checks/aria/aria-required-parent-evaluate.js
@@ -1,43 +1,33 @@
-import { implicitNodes, requiredContext } from '../../commons/aria';
-import { findUpVirtual, getRootNode } from '../../commons/dom';
-import {
-	getNodeFromTree,
-	matchesSelector,
-	escapeSelector
-} from '../../core/utils';
-
-function getSelector(role) {
-	var impliedNative = implicitNodes(role) || [];
-	return impliedNative.concat('[role="' + role + '"]').join(',');
-}
+import { getExplicitRole, getRole, requiredContext } from '../../commons/aria';
+import { getRootNode } from '../../commons/dom';
+import { getNodeFromTree, escapeSelector } from '../../core/utils';
 
 function getMissingContext(virtualNode, reqContext, includeElement) {
-	var index,
-		length,
-		role = virtualNode.actualNode.getAttribute('role'),
-		missing = [];
+	const missing = [];
+	const explicitRole = getExplicitRole(virtualNode);
 
 	if (!reqContext) {
-		reqContext = requiredContext(role);
+		reqContext = requiredContext(explicitRole);
 	}
 
 	if (!reqContext) {
 		return null;
 	}
 
-	for (index = 0, length = reqContext.length; index < length; index++) {
-		if (
-			includeElement &&
-			matchesSelector(virtualNode.actualNode, getSelector(reqContext[index]))
-		) {
-			return null;
+	for (let index = 0; index < reqContext.length; index++) {
+		let vNode = includeElement ? virtualNode : virtualNode.parent;
+
+		while (vNode) {
+			const parentRole = getRole(vNode);
+
+			if (reqContext[index].includes(parentRole)) {
+				return null;
+			}
+
+			vNode = vNode.parent;
 		}
-		if (findUpVirtual(virtualNode, getSelector(reqContext[index]))) {
-			//if one matches, it passes
-			return null;
-		} else {
-			missing.push(reqContext[index]);
-		}
+
+		missing.push(reqContext[index]);
 	}
 
 	return missing;
