+++ lib/commons/aria/implicit-nodes.js
@@ -5,6 +5,7 @@ import lookupTable from './lookup-table';
  * Get a list of CSS selectors of nodes that have an implicit role
  * @method implicitNodes
  * @memberof axe.commons.aria
+ * @deprecated
  * @instance
  * @param {String} role The role to check
  * @return {Mixed} Either an Array of CSS selectors or `null` if there are none
