+++ lib/commons/aria/required-context.js
@@ -1,4 +1,4 @@
-import lookupTable from './lookup-table';
+import standards from '../../standards';
 
 /**
  * Get the required context (parent) roles for a given role
@@ -9,15 +9,13 @@ import lookupTable from './lookup-table';
  * @return {Mixed} Either an Array of required context elements or `null` if there are none
  */
 function requiredContext(role) {
-	'use strict';
-	let context = null;
-	const roles = lookupTable.role[role];
+	const roleDef = standards.ariaRoles[role];
 
-	if (roles) {
-		// TODO: es-module-utils.clone
-		context = axe.utils.clone(roles.context);
+	if (!roleDef || !roleDef.requiredContext) {
+		return null;
 	}
-	return context;
+
+	return roleDef.requiredContext;
 }
 
 export default requiredContext;
