+++ test/commons/aria/roles.js
@@ -54,35 +54,6 @@ describe('aria.requiredOwned', function() {
 	});
 });
 
-describe('aria.requiredContext', function() {
-	'use strict';
-
-	var orig;
-	beforeEach(function() {
-		orig = axe.commons.aria.lookupTable.role;
-	});
-
-	afterEach(function() {
-		axe.commons.aria.lookupTable.role = orig;
-	});
-
-	it('should returned the context property for the proper role', function() {
-		axe.commons.aria.lookupTable.role = {
-			cats: {
-				context: 'yes'
-			}
-		};
-		assert.equal(axe.commons.aria.requiredContext('cats'), 'yes');
-	});
-
-	it('should return null if there are no required context nodes', function() {
-		axe.commons.aria.lookupTable.role = {};
-		var result = axe.commons.aria.requiredContext('cats');
-
-		assert.isNull(result);
-	});
-});
-
 describe('aria.implicitNodes', function() {
 	'use strict';
